+++ trixnity-messenger-compose-view/src/webMain/kotlin/de/connect2x/messenger/compose/view/startMessenger.kt
@@ -4,6 +4,9 @@ import androidx.compose.runtime.CompositionLocalProvider
 import androidx.compose.runtime.collectAsState
 import androidx.compose.ui.ExperimentalComposeUiApi
 import androidx.compose.ui.window.CanvasBasedWindow
+import com.arkivanov.decompose.DefaultComponentContext
+import com.arkivanov.essenty.lifecycle.Lifecycle
+import com.arkivanov.essenty.lifecycle.LifecycleRegistry
 import de.connect2x.messenger.compose.view.theme.MessengerTheme
 import de.connect2x.trixnity.messenger.MatrixMessengerConfiguration
 import de.connect2x.trixnity.messenger.createRoot
@@ -14,9 +17,16 @@ import de.connect2x.trixnity.messenger.multi.singleMode
 import io.github.oshai.kotlinlogging.KotlinLogging
 import io.github.oshai.kotlinlogging.KotlinLoggingConfiguration.logLevel
 import io.github.oshai.kotlinlogging.Level
-import kotlinx.browser.window
 import kotlinx.coroutines.flow.MutableStateFlow
+import kotlinx.coroutines.flow.updateAndGet
 import org.jetbrains.skiko.wasm.onWasmReady
+import web.dom.DocumentVisibilityState
+import web.dom.document
+import web.dom.visibilityChange
+import web.events.Event
+import web.events.addEventListener
+import web.uievents.FocusEvent
+import web.window.window
 
 private val log = KotlinLogging.logger {}
 
@@ -26,24 +36,46 @@ suspend fun startMessenger(
 ) {
     log.info { "Starting client" }
     logLevel = Level.DEBUG
-    val windowIsFocused = MutableStateFlow(false)
-    window.onfocus = {
-        log.debug { "window is focused" }
-        windowIsFocused.value = true
-        Unit
-    }
-    window.onblur = {
-        log.debug { "window is blurred" }
-        windowIsFocused.value = false
-        Unit
-    }
 
     val matrixMultiMessenger = MatrixMultiMessenger.create(configuration = configuration)
+    val lifecycleRegistry = LifecycleRegistry(Lifecycle.State.STARTED)
+    val windowIsFocused = MutableStateFlow(false)
 
     log.info { "Created MatrixMultiMessenger" }
+
+    document.addEventListener(
+        type = Event.visibilityChange(),
+        handler = { _: Event ->
+            lifecycleRegistry.updateState(
+                document.visibilityState == DocumentVisibilityState.visible,
+                windowIsFocused.value
+            )
+        }
+    )
+
+    window.addEventListener(
+        type = FocusEvent.focus(),
+        handler = { _: Event ->
+            lifecycleRegistry.updateState(
+                visible = document.visibilityState == DocumentVisibilityState.visible,
+                focused = windowIsFocused.updateAndGet { true }
+            )
+        }
+    )
+
+    window.addEventListener(
+        type = FocusEvent.blur(),
+        handler = { _: Event ->
+            lifecycleRegistry.updateState(
+                visible = document.visibilityState == DocumentVisibilityState.visible,
+                focused = windowIsFocused.updateAndGet { false }
+            )
+        }
+    )
+
     matrixMultiMessenger.singleMode { matrixMessenger ->
         try {
-            val rootViewModel = matrixMessenger.createRoot()
+            val rootViewModel = matrixMessenger.createRoot(DefaultComponentContext(lifecycleRegistry))
             val config = matrixMessenger.di.get<MatrixMessengerConfiguration>()
             onWasmReady {
                 CanvasBasedWindow(config.appName) {
@@ -74,6 +106,29 @@ suspend fun startMessenger(
             println("-------")
             println(e.stackTraceToString())
             throw e
+        }
+    }
+}
+
+private fun LifecycleRegistry.updateState(visible: Boolean, focused: Boolean) {
+    val target = when {
+        visible && focused -> Lifecycle.State.RESUMED
+        visible -> Lifecycle.State.STARTED
+        else -> Lifecycle.State.CREATED
+    }
+    if (state != target) {
+        log.debug { "Application State changing from $state to $target" }
+        while (state < target) when (state) {
+            Lifecycle.State.INITIALIZED -> onCreate()
+            Lifecycle.State.CREATED -> onStart()
+            Lifecycle.State.STARTED -> onResume()
+            else -> Unit
+        }
+        while (state > target) when (state) {
+            Lifecycle.State.RESUMED -> onPause()
+            Lifecycle.State.STARTED -> onStop()
+            Lifecycle.State.CREATED -> onDestroy()
+            else -> Unit
         }
     }
 }